+++ app/actions/accountActions.js
@@ -6,9 +6,6 @@ import claimsActions from './claimsActions'
 import pricesActions from './pricesActions'
 import transactionHistoryActions from './transactionHistoryActions'
 import { getPendingTransactionInfo } from './pendingTransactionActions'
-// import { blockHeightActions } from './blockHeightActions'
-import nftActions from './nftActions'
-import nftGalleryActions from './nftGalleryActions'
 
 export const ID = 'account'
 
@@ -18,7 +15,4 @@ export default createBatchActions(ID, {
   transactions: transactionHistoryActions,
   pendingTransactionsInfo: getPendingTransactionInfo,
   prices: pricesActions,
-  // blockHeight: blockHeightActions,
-  nft: nftActions,
-  nftGallery: nftGalleryActions,
 })
