+++ app/components/ErrorBoundaries/Main/index.js
@@ -1,20 +1,12 @@
 // @flow
 import { compose } from 'recompose'
-import { withActions } from 'spunky'
 
 import Main from './Main'
-import { logoutActions } from '../../../actions/authActions'
+
 import withSettingsContext from '../../../hocs/withSettingsContext'
-
-type Props = {
-  logout: Function,
-}
-
-const mapActionsToProps = (actions): Props => ({
-  logout: () => actions.call(),
-})
+import withAuthData from '../../../hocs/withAuthData'
 
 export default compose(
-  withActions(logoutActions, mapActionsToProps),
+  withAuthData,
   withSettingsContext,
 )(Main)
