+++ app/containers/Dashboard/index.js
@@ -2,16 +2,12 @@
 import { connect, type MapStateToProps } from 'react-redux'
 import { bindActionCreators } from 'redux'
 import { compose } from 'recompose'
-import { withReset, withActions, withData } from 'spunky'
 
-import dashboardActions from '../../actions/dashboardActions'
-import accountActions from '../../actions/accountActions'
-import withInitialCall from '../../hocs/withInitialCall'
 import withAuthData from '../../hocs/withAuthData'
 import withNetworkData from '../../hocs/withNetworkData'
 import { getNotifications } from '../../modules/notifications'
 import { showModal } from '../../modules/modal'
-import { internetConnectionPromptPresented } from '../../actions/internetConnectivityPromptActions'
+
 import Dashboard from './Dashboard'
 import withSettingsContext from '../../hocs/withSettingsContext'
 
@@ -26,15 +22,15 @@ const actionCreators = {
 const mapDispatchToProps = dispatch =>
   bindActionCreators(actionCreators, dispatch)
 
-const mapAccountActionsToProps = (actions, props) => ({
-  loadWalletData: () =>
-    actions.call({
-      net: props.net,
-      address: props.address,
-      tokens: props.tokens,
-      chain: props.chain,
-    }),
-})
+// const mapAccountActionsToProps = (actions, props) => ({
+//   loadWalletData: () =>
+//     actions.call({
+//       net: props.net,
+//       address: props.address,
+//       tokens: props.tokens,
+//       chain: props.chain,
+//     }),
+// })
 
 export default compose(
   connect(
@@ -42,12 +38,12 @@ export default compose(
     mapDispatchToProps,
   ),
   withAuthData,
-
+  withSettingsContext,
   // Expose function for polling & reloading account related data.
   withNetworkData(),
-  withInitialCall(dashboardActions),
-  withReset(accountActions, ['networkId']),
-  withActions(accountActions, mapAccountActionsToProps),
-  withData(internetConnectionPromptPresented),
-  withAuthData,
-)(withSettingsContext(Dashboard))
+
+  // withReset(accountActions, ['networkId']),
+  // withActions(accountActions, mapAccountActionsToProps),
+  // withData(internetConnectionPromptPresented),
+  // withAuthData,
+)(Dashboard)
